Remove %f modifiers in debug print statements.
The "%f" modifier causes a problem to the emulator because internally,
when printf encounters that modifier, it uses dlmalloc to create some
space, presumably to hold the contents of that modifier. The space
allocated might be in a tainted page and as a result, when libc tries to
access that space, the emulator will encounter a repeated fault.